+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**3.2.0** (June 2, 2015)
+
+* Updated SRI integrity to align with spec changes
+* Deprecated Manifest integrity attribute
+* Cleanup concatenating JS sources with newlines
+
 **3.1.0** (May 10, 2015)
 
 * Removed "index" logical path normalization. Asset#logical_path is always the